Prepare for ThingSet Mesh: Add RBBQ Circular Buffer Library
The RBBQ libary provides a bi-directional message buffer for local and remote communication.
The messages are queued up in a bi-partite circular buffer. The circular buffer always
supports writing a contiguous message, instead of potentially splitting a write in two chunks
when it straddles the buffer's boundaries (similar to
`BipBuffer <https://www.codeproject.com/Articles/3479/The-Bip-Buffer-The-Circular-Buffer-with-a-Twist>`_).

The use case is communication between two concurrent threads of control: either two actual OS threads,
or two interconnected devices.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>